refactor(goals): neutralize goal runtime service naming
